Finish line

+++ src/main/java/njt/myproject/dax/services/TodoListService.java
@@ -1,6 +1,7 @@
 package njt.myproject.dax.services;
 
 import javassist.NotFoundException;
+import njt.myproject.dax.exceptions.PermissionDeniedException;
 import njt.myproject.dax.models.TodoList;
 import njt.myproject.dax.models.User;
 import njt.myproject.dax.models.UserHasTodoList;
@@ -43,10 +44,12 @@ public class TodoListService {
     }
 
     @Transactional
-    public TodoList updateTodoList(int id, String name, User user) throws NotFoundException {
+    public TodoList updateTodoList(int id, String name, User user) throws NotFoundException, PermissionDeniedException {
         TodoList todoList = todoListRepository.findById(id).orElse(null);
         if (todoList != null) {
-            //TODO premission check
+            // premission check
+            checkPermission(user.getId(),todoList.getId());
+
             todoList.setName(name);
             return todoListRepository.save(todoList);
         } else {
@@ -55,10 +58,12 @@ public class TodoListService {
     }
 
     @Transactional
-    public void deleteTodoList(int id, int user_id) throws NotFoundException {
+    public void deleteTodoList(int id, int user_id) throws NotFoundException, PermissionDeniedException {
         TodoList todoList = todoListRepository.findById(id).orElse(null);
         if (todoList != null) {
-            //TODO premission check
+            // premission check
+            checkPermission(user_id, id);
+
             todoListRepository.delete(todoList);
         } else {
             throw new NotFoundException("Todo list not found.");
@@ -69,4 +74,13 @@ public class TodoListService {
         return userHasTodoListRepository.getByUserId(user_id);
     }
 
+    private void checkPermission(int user_id, int task_list_id) throws PermissionDeniedException {
+        UserHasTodoListPK userHasTodoListPK = new UserHasTodoListPK();
+        userHasTodoListPK.setUserId(user_id);
+        userHasTodoListPK.setTodoListId(task_list_id);
+        if (userHasTodoListRepository.findById(userHasTodoListPK).orElse(null) == null) {
+            throw new PermissionDeniedException("You don't have premmission for this");
+        }
+
+    }
 }
